# share/src/main/resources/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "REST API SPEC FOR MOVESONG FRONTEND",
    "version" : "v1",
    "title" : "REST API SPEC FOR MOVESONG FRONTEND"
  },
  "host" : "localhost:8060",
  "basePath" : "/share",
  "tags" : [ {
    "name" : "share"
  } ],
  "schemes" : [ "http" ],
  "paths" : {
    "/createShare" : {
      "post" : {
        "tags" : [ "share" ],
        "summary" : "Create a share",
        "description" : "",
        "operationId" : "createShare",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/CreateShareReq"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/CreateShareResp"
            }
          }
        }
      }
    },
    "/getShareById" : {
      "post" : {
        "tags" : [ "share" ],
        "summary" : "Get a share by the id",
        "description" : "",
        "operationId" : "getShareById",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/GetShareByIdReq"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/GetShareByIdResp"
            }
          }
        }
      }
    },
    "/getSharesByMovesongEmail" : {
      "post" : {
        "tags" : [ "share" ],
        "summary" : "Get all shares by the movesong email",
        "description" : "",
        "operationId" : "getSharesByMovesongEmail",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/GetSharesByMovesongEmailReq"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/GetSharesByMovesongEmailResp"
            }
          }
        }
      }
    },
    "/updateShare" : {
      "post" : {
        "tags" : [ "share" ],
        "summary" : "Update a share",
        "description" : "",
        "operationId" : "updateShare",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/UpdateShareReq"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/UpdateShareResp"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "CreateShareReq" : {
      "type" : "object",
      "properties" : {
        "share" : {
          "$ref" : "#/definitions/ShareDto"
        }
      }
    },
    "CreateShareResp" : {
      "type" : "object",
      "properties" : {
        "share" : {
          "$ref" : "#/definitions/ShareDto"
        }
      }
    },
    "GetShareByIdReq" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "GetShareByIdResp" : {
      "type" : "object",
      "properties" : {
        "share" : {
          "$ref" : "#/definitions/ShareDto"
        }
      }
    },
    "GetSharesByMovesongEmailReq" : {
      "type" : "object",
      "properties" : {
        "movesongEmail" : {
          "type" : "string"
        }
      }
    },
    "GetSharesByMovesongEmailResp" : {
      "type" : "object",
      "properties" : {
        "shares" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ShareDto"
          }
        }
      }
    },
    "ShareDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "playlistId" : {
          "type" : "string"
        },
        "sharedPlaylistName" : {
          "type" : "string"
        },
        "ownerMovesongEmail" : {
          "type" : "string"
        },
        "visible" : {
          "type" : "boolean"
        },
        "views" : {
          "type" : "integer",
          "format" : "int32"
        },
        "sharePlatformType" : {
          "type" : "string"
        },
        "selectedBackgroundId" : {
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "UpdateShareReq" : {
      "type" : "object",
      "properties" : {
        "share" : {
          "$ref" : "#/definitions/ShareDto"
        }
      }
    },
    "UpdateShareResp" : {
      "type" : "object",
      "properties" : {
        "share" : {
          "$ref" : "#/definitions/ShareDto"
        }
      }
    }
  }
}